requirements.txt
pandas>=2.0
numpy>=1.24
deap>=1.3.3
optuna>=3.4
joblib>=1.3
pydantic>=2.5
PyYAML>=6.0.1
tqdm>=4.66
requests>=2.32
SQLAlchemy>=2.0
numba>=0.59; platform_system != "Windows"  # optional, ignore on Windows
